New library function to compute study profile
Display a table representing the study profile of a concrete area and inconsistencies among numbers.

New library function called StudyProfileOfAre which computes a table using kable representing the study profile of the specified study area.

New auxiliar function called Pivot to convert a data frame of two columns in a list in which once column is used as keys and the other as values.